pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "diffmjx"
version = "0.0.1"
description = "Adaptive time integration for MuJoCo MJX using diffrax"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.12"
dependencies = [
    "jax>=0.4.20",
    "equinox>=0.11.0",
    "optimistix>=0.0.6",
    "softjax",
    "mujoco==3.3.1",
    "mujoco-mjx==3.3.1",
    "diffrax>=0.6.0",
    "mjx-diffrax",
    "matplotlib",
    "seaborn",
    "hydra-core",
    "omegaconf",
    "wandb",
    "mediapy",
    "tqdm",
]

[project.optional-dependencies]
gpu = [
  "jax[cuda12]>=0.8.0",
]
dev = ["pytest", "ruff"]

[tool.uv.sources]
softjax = { path = "./external/softjax", editable = true }
mujoco-mjx = { path = "./external/mujoco/mjx", editable = true }
mjx-diffrax = { path = "./external/mjx_diffrax", editable = true }

[tool.setuptools]
packages = []

[tool.ruff]
line-length = 100
extend-exclude = ["external"]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
]
